build search index during build time, add docs body to the search
